Made some get and set methods in level and stuff to make the above work

+++ src/com/sinius15/suite/game/Rail.java
@@ -2,7 +2,7 @@ package com.sinius15.suite.game;
 
 public abstract class Rail extends Tile{
 
-	public Rail(byte id) {
+	public Rail(int id) {
 		super(id);
 	}
